+++ global-service/src/main/java/in/cdac/university/globalService/entity/GbltConfigApplicationTrackerDtl.java
@@ -122,10 +122,10 @@ public class GbltConfigApplicationTrackerDtl implements Serializable {
 	private Long unumNdtlCourseId;
 
 	@Column(name="unum_ndtl_department_id")
-	private Long unumNdtlDepartmentId;
+	private Integer unumNdtlDepartmentId;
 
 	@Column(name="unum_ndtl_faculty_id")
-	private Long unumNdtlFacultyId;
+	private Integer unumNdtlFacultyId;
 
 	@Column(name="unum_out_no")
 	private String unumOutNo;
